WGI: Use polling to detect new gamepads
The add/remove callbacks that WGI provides don't integrate well with SDL's
joystick usage model, because they can arrive after WGI_JoystickQuit() on
arbitrary threads. This can lead to races with the JoystickQuit() invocation
of DInput, HIDAPI, or even WGI itself.

Even if we properly refcounted our WGI state to handle those delayed callbacks,
it would be very challenging to correctly synchronize with the state of other
joystick implementations that may be destroying/destroyed at the time of the
callback.

Simply use the add/remove callbacks as an indication to redetect when
WGI_JoystickDetect() is called again.